add new component GridContainer(9 news grid)
