+++ Navbar.css
@@ -6,7 +6,7 @@
 	display: inline-grid;
 	grid-template-rows: 1fr 4fr 1fr 1fr;
 	height: 100vh;
-	width: 4vw;
+	width: 3.6vw;
 	
 	color: white;
 	background-color: #808080;
