+++ index.html
@@ -9,8 +9,10 @@
 	<style type="text/css">
 		body{
 			margin: 0px;
+			width: 600vw;/*//////////////////////////////////////////*/
 			height: 100vh;
 			background-color: #f0f0f0;
+			overflow-y: hidden;
 		}
 		#root{
 			margin: 0px;
